test: add unit tests and fix _find_split_point float index bug
- 90 tests across 4 test files: utils, ingestion, core_memory, db
- All DB helper tests use mocked connections (no DB required)
- Fix _find_split_point float slice index bug in ingestion.py
  (max_len * 0.5 produced float, rfind requires int)

+++ backend/src/agent/storage/ingestion.py
@@ -142,7 +142,7 @@ def _find_split_point(text: str, max_len: int) -> int:
             if best >= max_len * 0.5:
                 return best
     # Fallback: split at space
-    space = text.rfind(" ", max_len * 0.5, max_len)
+    space = text.rfind(" ", int(max_len * 0.5), max_len)
     if space > 0:
         return space
     return max_len
